2010-04-14	Ralf Corsépius <ralf.corsepius@rtems.org>
	* libnetworking/libc/gethostbydns.c: Use uintptr_t instead of u_long
	for 16bit target compliance.

+++ cpukit/libnetworking/libc/gethostbydns.c
@@ -393,7 +393,7 @@ gethostanswer(
 				buflen -= nn;
 			}
 
-			bp += sizeof(align) - ((u_long)bp % sizeof(align));
+			bp += sizeof(align) - ((uintptr_t)bp % sizeof(align));
 
 			if (bp + n >= &hostbuf[sizeof hostbuf]) {
 				debugprintf("size (%d) too big\n", n);
